requirements.txt
psutil>=5.9.0
watchdog>=2.1.0
python-magic>=0.4.24
cachetools>=5.0.0
concurrent-futures>=3.1.1
pillow>=9.0.0
pandas>=1.5.0
openpyxl>=3.0.0
reportlab>=3.6.0
Pillow>=9.0.0
cryptography>=3.4.0
requests>=2.28.0
streamlit>=1.20.0